gaiguan-mall/requirement.txt
aiohttp==3.6.2
amqp==2.5.2
appdirs==1.4.3
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
Automat==0.7.0
beautifulsoup4==4.8.0
billiard==3.5.0.5
bs4==0.0.1
celery==4.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
defusedxml==0.6.0
distlib==0.3.0
Django==2.2.1
django-allauth==0.39.1
django-ckeditor==5.6.1
django-crispy-forms==1.8.1
django-haystack==2.8.1
django-jet==1.0.8
django-js-asset==1.2.2
django-model-utils==4.0.0
django-mptt==0.10.0
django-notifications-hq==1.5.0
django-password-reset==2.0
django-ranged-response==0.2.0
django-redis==4.11.0
django-simple-captcha==0.5.12
django-taggit==0.23.0
django-tinymce==2.8.0
dnspython==1.16.0
docopt==0.6.2
easygui==0.98.1
eventlet==0.25.1
fake-useragent==0.1.11
filelock==3.0.12
get==2019.4.13
greenlet==0.4.15
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==17.5.0
itchat==1.3.10
itsdangerous==1.1.0
jieba==0.42.1
jsonfield==2.0.2
kombu==4.6.8
lxml==4.4.1
Markdown==3.1.1
monotonic==1.5
multidict==4.7.5
mysql-connector-python==8.0.19
mysqlclient==1.4.6
Naked==0.1.31
numpy==1.17.2
oauthlib==3.1.0
opencv-python==4.1.1.26
parsel==1.5.2
Pillow==5.1.0
pipreqs==0.4.10
post==2019.4.13
public==2019.4.13
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
pycryptodomex==3.9.4
PyDispatcher==2.0.5
pygame==1.9.6
PyHamcrest==1.9.0
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pypng==0.0.20
PyQRCode==1.2.1
PyQt5Designer==5.14.1
python-alipay-sdk==2.0.1
python-docx==0.8.10
python3-openid==3.1.0
pytz==2019.3
pywin32==221
PyYAML==5.3.1
qrcode==6.1
query-string==2019.4.13
queuelib==1.5.0
redis==3.4.1
request==2019.4.13
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
Scrapy==1.7.3
selenium==3.141.0
service-identity==18.1.0
shellescape==3.8.1
six==1.12.0
slugify==0.0.1
soupsieve==1.9.3
sqlparse==0.3.1
Twisted==19.7.0
urllib3==1.25.3
vine==1.3.0
virtualenv==20.0.10
w3lib==1.21.0
Whoosh==2.7.4
yarg==0.1.9
yarl==1.4.2
zipp==3.1.0
zope.interface==4.6.0
